Restructured metaclasses and added versionedinitmeta as mixin. Added automatic version finding when the __new__ is called from the original VersionedClass definition.
